+++ src/ansys/systemcoupling/core/session.py
@@ -1,7 +1,7 @@
 from typing import Callable, Optional
 
+from ansys.systemcoupling.core.adaptor.impl.source import get_root
 from ansys.systemcoupling.core.adaptor.impl.syc_proxy import SycProxy
-from ansys.systemcoupling.core.adaptor.impl.types import Container, get_root
 from ansys.systemcoupling.core.native_api import NativeApi
 
 try:
@@ -14,6 +14,8 @@ try:
 except ImportError:
     # Fallback to generic type
     # (should not occur in normal use and in doc generation)
+    from ansys.systemcoupling.core.adaptor.impl.types import Container
+
     case_root = setup_root = solution_root = Container
